first commit.
Got this code off of Google App Engine painstakingly when they
announced that Python 2.5 was no longer supported, and I must
upgrade this project to Python 2.7 to continue using it.
Screw that.  I couldn't even download the code.  I had to resort
to copying and pasting off of their online source viewer.

Keeping this for old times' sake.